% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_df}
\alias{aggregate_df}
\title{Aggregates matrix or data frame}
\usage{
aggregate_df(df, groups, fun = mean, margin = 1, ...)
}
\arguments{
\item{df}{A matrix or data frame}

\item{groups}{A vector of groups (discrete values)}

\item{fun}{A function that aggregates a vector into objects of length 1}

\item{margin}{If 1, aggregates rows, if 2 aggregates columns. Defaults to 1}

\item{...}{passed to fun}
}
\value{
A data.frame with aggregated rows or columns
}
\description{
Aggregates matrix or data frame
}
