% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorme.R
\name{colorme}
\alias{colorme}
\title{colorme}
\usage{
colorme(col = NULL, numb = NULL)
}
\arguments{
\item{col}{Name of color.}

\item{numb}{Number of colors to return}
}
\description{
Returns colors from EB color scheme
}
\examples{
colorme("blue")
}
\keyword{file}

