% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workspace_and_import_files.R
\name{create_namesfile_template}
\alias{create_namesfile_template}
\title{Create a 'names_template.csv' file from fcs files in the /fcs/config folder}
\usage{
create_namesfile_template(fs = fs_src,
  path = "./config/names_template.csv")
}
\description{
Create a 'names_template.csv' file from fcs files in the /fcs/config folder
}
\details{
The template can be copied, edited to select relevant variables, and saved to be reused in future functions such as tSNE, ONESENSE etc...
}
\examples{
create_namesfile_template()
}
