% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_outputs.R
\name{data_ranges}
\alias{data_ranges}
\title{Computes data ranges for each channel of each flowFrame of a flowSet}
\usage{
data_ranges(fs, global_across_samples, global_across_channels, na.rm = T)
}
\arguments{
\item{fs}{A flowSet}

\item{global_across_channels}{Boolean}

\item{na.rm}{Boolean : wether to ignore missing values or change output to NA}
}
\description{
Computes data ranges for each channel of each flowFrame of a flowSet
}
\details{
See \code{\link{color_biplots_by_channels}} for details of how the boolean arguments affect the output
}
