% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{en.FlowSOM}
\alias{en.FlowSOM}
\title{FlowSOM wrapper to call on matrices}
\usage{
en.FlowSOM(mat, scale = FALSE, scaled.center = TRUE,
  scaled.scale = TRUE, nClus = NULL, maxMeta, ...)
}
\description{
FlowSOM wrapper to call on matrices
}
