% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{en.load}
\alias{en.load}
\title{Allows to bind the output of load() to a variable}
\usage{
en.load(what)
}
\arguments{
\item{what}{character, filename passed to load}
}
\description{
Allows to bind the output of load() to a variable
}
