% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workspace_and_import_files.R
\name{en.read_flowset}
\alias{en.read_flowset}
\title{Reads a flowset and converts to a list with 1) a big expression matrix with targets as colnames 2) a named vector of matrix rownames to samples 3) a named vector of colnames to channels}
\usage{
en.read_flowset(files, filter = TRUE, compensate = FALSE, ...)
}
\description{
Reads a flowset and converts to a list with 1) a big expression matrix with targets as colnames 2) a named vector of matrix rownames to samples 3) a named vector of colnames to channels
}
