% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_or_export_workspace_content.R
\name{en.write_flowset}
\alias{en.write_flowset}
\title{Wrapper to flowCore:::write.flowSet to preserve channel annotation}
\usage{
en.write_flowset(fs, directory)
}
\arguments{
\item{fs}{A flowSet}

\item{directory}{Where to save the flowFrames of the flowSet}
}
\description{
Wrapper to flowCore:::write.flowSet to preserve channel annotation
}
