% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_outputs.R
\name{fastBreaks}
\alias{fastBreaks}
\title{Computes breaks for a color scale based on a matrix}
\usage{
fastBreaks(Mat, n, sym = T)
}
\arguments{
\item{Mat}{a matrix}

\item{n}{How many breaks}

\item{sym}{Boolean. Whether breaks should be symetric (centered at 0)}
}
\value{
Numerical vector of length n
}
\description{
Computes breaks for a color scale based on a matrix
}
