% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_workspace_and_import_files.R
\name{filter_fcs}
\alias{filter_fcs}
\title{Use regular expressions to only filter files ending in .fcs from a vector of files}
\usage{
filter_fcs(files)
}
\arguments{
\item{files}{Files to filter}
}
\value{
A character vector of .fcs files
}
\description{
Use regular expressions to only filter files ending in .fcs from a vector of files
}
