% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freqplot}
\alias{freqplot}
\title{Plot a 2D heatmap of frequencies using rectangular bins (a la FlowJo)}
\usage{
freqplot(x, y, breaks = 200, na.rm = TRUE, ...)
}
\description{
Plot a 2D heatmap of frequencies using rectangular bins (a la FlowJo)
}
