% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting flowsets.R
\name{gate_from_biplot}
\alias{gate_from_biplot}
\title{From a biplot let the user interactively draw polygons to create a "Gate" column in the expression data}
\usage{
gate_from_biplot(matrix, x_axis, y_axis, ..., bty = "l", pch = 16,
  cex = 0.5, alpha = 100, sample = NULL)
}
\arguments{
\item{matrix}{A matrix}

\item{...}{passed to plot}

\item{alpha}{Transparency value (between 0 and 255, 0 is fully transparent)}
}
\value{
A named vector of length nrow(matrix) and names rownames(matrix). Ungated events are set to NA
}
\description{
From a biplot let the user interactively draw polygons to create a "Gate" column in the expression data
}
