% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_outputs.R
\name{imageWrapper}
\alias{imageWrapper}
\title{Wrapper to image where the coordinate system is more intuitive}
\usage{
imageWrapper(
  matrixToPlot,
  labCol = colnames(matrixToPlot),
  labRow = rownames(matrixToPlot),
  ...,
  cex.text = 1,
  cex.row = 1,
  cex.col = 1,
  na.col = par("bg")
)
}
\arguments{
\item{matrixToPlot}{A matrix with two dimensions}

\item{labCol}{if specified, the text to annotate rows}

\item{labRow}{if specified, the text to annotate colums}

\item{...}{passed to image}

\item{cex.text}{cex value for both rows and columns}

\item{cex.row}{cex value for rows only. overrides cex.text}

\item{cex.col}{cex value for columns only. overrides cex.text}

\item{na.col}{color for missing values}
}
\description{
Wrapper to image where the coordinate system is more intuitive
}
