% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{line2user}
\alias{line2user}
\title{Convert line to user coordinates}
\usage{
line2user(line, side)
}
\description{
Convert line to user coordinates
}
\author{
Credit to jbaums https://stackoverflow.com/questions/30765866/get-margin-line-locations-in-log-space/30835971#30835971
}
