% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_projects.R
\name{make_card_project}
\alias{make_card_project}
\title{Create a project card (for Distill website)}
\usage{
make_card_project(
  title = NULL,
  title_link = NULL,
  description = NULL,
  button = "See more",
  button_link = NULL
)
}
\arguments{
\item{button_link}{}
}
\value{
HTML for a card
}
\description{
Create a project card (for Distill website)
}
