% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_projects.R
\name{make_grid_project}
\alias{make_grid_project}
\title{Create a grid with projects cards (for Distill website)}
\usage{
make_grid_project(...)
}
\arguments{
\item{...}{Cards to include with make_project_card}
}
\value{
HTML to create a grid
}
\description{
Create a grid with projects cards (for Distill website)
}
