% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Makes a color transparent}
\usage{
makeTransparent(colors, alpha = 255)
}
\arguments{
\item{colors}{A vector of colors as in `?col2rgb`}

\item{alpha}{transparency value (0=fully transparent, 255=fully opaque)}
}
\description{
Makes a color transparent
}
