% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{month_to_number}
\alias{month_to_number}
\title{Convert month name to month number}
\usage{
month_to_number(month_name)
}
\arguments{
\item{month_name}{Month name (partial match authorized)}
}
\value{
An integer
}
\description{
Convert month name to month number
}
