% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting flowsets.R
\name{namefile_to_colnames}
\alias{namefile_to_colnames}
\title{Takes from input a channelsname.csv file and returns columns to select as a vector}
\usage{
namefile_to_colnames(channels.namefile, to_config = TRUE)
}
\arguments{
\item{channels.namefile}{A csv file}

\item{to_config}{TRUE if file is in ./config/ subfolder, FALSE if it's just a file path}
}
\value{
a character vector of column names to be selected
}
\description{
Takes from input a channelsname.csv file and returns columns to select as a vector
}
