% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecd.R
\name{oecd_countries}
\alias{oecd_countries}
\title{Print the R code to make the vector of 38 OECD countries used in Beine et al. 2020}
\usage{
oecd_countries()
}
\description{
Print the R code to make the vector of 38 OECD countries used in Beine et al. 2020
}
