% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prior_free_bayes_call}
\alias{prior_free_bayes_call}
\title{Bayes classifier - prediction}
\usage{
prior_free_bayes_call(data, model)
}
\description{
Bayes classifier - prediction
}
\note{
To incorporate a class-frequency prior, multiply each score by the class frequency on the training data
}
