% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_matrix}
\alias{split_matrix}
\title{Fast splitting of matrix to list (avoids conversion do data.frame)}
\usage{
split_matrix(mat, vector, byrow = TRUE)
}
\description{
Fast splitting of matrix to list (avoids conversion do data.frame)
}
