% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toColors_continuous}
\alias{toColors_continuous}
\title{Map a vector to colors (for continuous variables); new version}
\usage{
toColors_continuous(vector, palette = rev(c("#A50026", "#D73027",
  "#F46D43", "#FDAE61", "#FEE090", "#FFFFBF", "#E0F3F8", "#ABD9E9",
  "#74ADD1", "#4575B4", "#313695")))
}
\description{
Map a vector to colors (for continuous variables); new version
}
