% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toColors_discrete}
\alias{toColors_discrete}
\title{Map a vector to colors (for categorical variables)}
\usage{
toColors_discrete(
  vector,
  palette = c("#8DD3C7", "#FFFFB3", "#BEBADA", "#FB8072", "#80B1D3", "#FDB462",
    "#B3DE69", "#FCCDE5", "#D9D9D9", "#BC80BD", "#CCEBC5", "#FFED6F")
)
}
\description{
Map a vector to colors (for categorical variables)
}
