% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhereAmI.R
\name{WhereAmI}
\alias{WhereAmI}
\title{WhereAmI}
\usage{
WhereAmI(fn)
}
\arguments{
\item{fn}{Name of the file you want to locate.}
}
\description{
Allows to locate a file in Mac OS
}
\examples{
WhereAmI("FSS-Pathfinder.R")
}
\keyword{file}

