% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeClipboard}
\alias{writeClipboard}
\title{writeClipboard for Linux}
\usage{
writeClipboard(x, sep = "\\t", row.names = FALSE, col.names = TRUE)
}
\description{
writeClipboard for Linux
}
\author{
Credit to Freecube on http://stackoverflow.com/questions/10959521/how-to-write-to-clipboard-on-ubuntu-linux-in-r
}
